[DAGE-3] Renamed search engine interface
